+++ hal/cleanup/TODO
@@ -1,0 +1,5 @@
+a.c b.c. c.c. d.c need to be compressed into one XMACRO
+
+the four files must pasted into stitch.c
+
+a single xmacro (xmacro.c) must expand the an match stitch.c
